server/requirements.txt
# Python 3.12+ Compatible Requirements
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.28
email-validator>=2.0.0
